Change time_t to int64_t

+++ kernel/posix/sys/types.h
@@ -32,7 +32,7 @@ typedef uint32_t dev_t;
 typedef uint16_t mode_t;
 typedef uint32_t nlink_t;
 
-typedef int32_t time_t;
+typedef int64_t time_t;
 typedef uint32_t useconds_t;
 typedef int32_t suseconds_t;
 typedef uint64_t clock_t;
